change some numberics to Long type

+++ src/main/java/io/jafka/jeos/core/response/chain/Block.java
@@ -13,7 +13,7 @@ public class Block {
 
     private String actionMroot;
     private Long blockNum;
-    private Integer confirmed;
+    private Long confirmed;
     private String id;
     private String previous;
     private String producer;
